auth_service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
pydantic>=2.5.0
python-multipart>=0.0.6 # For OAuth2PasswordRequestForm / form data
email-validator>=2.1.0 # For EmailStr validation
httpx>=0.25.2 # For inter-service communication
psycopg2-binary>=2.9.7 # PostgreSQL adapter